add redis; add&config slf4j-logger; and some refactoring

+++ src/main/java/net/arxemond/bookmanager/dao/BookDao.java
@@ -1,18 +1,21 @@
 package net.arxemond.bookmanager.dao;
 
 import net.arxemond.bookmanager.model.Book;
-
+import org.springframework.cache.annotation.*;
 import java.util.List;
 
 public interface BookDao
 {
     public void addBook(Book book);
 
+//    @CachePut("book")
     public void updateBook(Book book);
 
-    public void removeBook(int id);
+//    @CacheEvict("book")
+    public void removeBook(Long id);
 
-    public Book getBookById(int id);
+//    @Cacheable("book")
+    public Book getBookById(Long id);
 
     public List<Book> listBooks();
 }
